feat: multi generator support in boxes_main and extend testing capabilities
The current test setup can only verify that generators default output
doesn't change; this change uses a YAML file to automate generation of
examples and supports making custom examples with custom arguments.

This is also reusable to automate creation of multiple boxes.  For
example, you are generating a set of boxes for a silverware drawer
and you can write a single YAML file and then share that file with
others to generate all the boxes.